Separate and clean structure / sound loaders, mostly remove CSVHelper
